rag_demo/requirements.txt
streamlit>=1.35.0
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
openai>=1.2.4
tqdm>=4.66.0
numpy>=1.24.0
PyMuPDF>=1.22.0
